The basics of investigation
When you click the circles, a dialogue will be triggered and the other circles hidden. The dialogue works similarly to that of when speaking with other characters. Next step would be to show an enlarged image of the object in the middle and add the evidence/contact when the inventory/phone mechanics are finished.