Audios are done (mostly)
For each new scenes, make sure that there's an audio manager (exists a prefab already) and there is audio source of "Typing_Two (use this)" in dialogue manager as well as its script contains the code for typing audio. Also, there are two different bgms to choose from (click on/off the play on awake and loop).